requirements.txt
asgiref==3.8.1
cloudinary==1.41.0
Django==5.1.4
django-cloudinary-storage==0.3.0
djangorestframework==3.15.2
pillow==11.1.0
pytz==2024.2
sqlparse==0.5.3
